refactor(optmizers): move optimizers into their own files, into package
