+++ images_optimization/optimizer.py
@@ -1,17 +1,12 @@
 import glob
 
-from PIL import Image
-
-from images_optimization import mozjpeg
+from images_optimization.optimizers import mozjpeg, pillow
 from images_optimization.logger import logger
 from images_optimization.args_parser import parse_args
 from images_optimization.permissions_fixer import fix_permissions
 
 jpg_extensions = ['jpg', 'jpeg']
 extensions = jpg_extensions + ['png', 'gif']
-
-max_image_width = 3000
-max_image_height = 3000
 
 
 def optimize_images_from_directory(directory):
@@ -67,9 +62,7 @@ def _optimize_non_jpeg_image(filename):
     logger.info('Optimizing « %s » with Pillow...' % filename)
 
     try:
-        image = Image.open(filename)
-        image.thumbnail((max_image_width, max_image_height), Image.ANTIALIAS)
-        image.save(filename, quality=80, optimize=True)
+        pillow.optimize(filename)
         fix_permissions(filename)
     except Exception:
         logger.exception('An error occurred during optimization with Pillow')
