+++ images_optimization/optimizers/mozjpeg.py
